improve handling of visibility

+++ api/views.py
@@ -1,5 +1,6 @@
 from django.core import serializers
-from django.http import HttpResponse, JsonResponse
+from django.http import HttpResponse, JsonResponse, HttpResponseNotFound
+from django.shortcuts import get_object_or_404
 from rest_framework.renderers import JSONRenderer
 from .serializers import PostSerializer, CommentSerializer, UserSerializer
 from django.shortcuts import render, redirect
@@ -7,7 +8,7 @@ from .forms import UserForm
 from .models import User, Post, Comment, Friend
 from django.core.paginator import Paginator
 from .utils import *
-from .filters import apply_filter
+from .filters import get_posts_by_status, get_public_posts, user_is_authorized
 import json
 
 # TODO: serializers should only spit out certain fields (per example-article.json), ez but tedious
@@ -28,8 +29,16 @@ def posts_visible(request):
     method = request.method
     if method == "GET":
         page, size, filter_ = get_post_query_params(request)
-        # TODO: posts visible to the currently authenticated user
-        posts = apply_filter(request, filter_)
+
+        if filter_:
+            posts = get_posts_by_status(filter_)
+        else:
+            posts = get_public_posts()
+
+        posts = list(filter(
+                lambda post: user_is_authorized(request.user, post),
+                posts))
+
         posts_pages = Paginator(posts, size)
         response_body = {
             "query": "posts",
@@ -59,9 +68,13 @@ def posts_by_aid(request, aid):
     if method == "GET":
         page, size, filter_ = get_post_query_params(request)
         # all posts made by author id aid visible to currently authed user
-        # TODO: this thing here ignores the visibility thing
         user = User.objects.get(pk=aid)
         posts = Post.objects.filter(author=user)
+
+        posts = list(filter(
+                lambda post: user_is_authorized(request.user, post),
+                posts))
+
         posts_pages = Paginator(posts, size)
         response_body = {
             "query": "posts",
@@ -78,9 +91,11 @@ def posts_by_aid(request, aid):
 def all_posts(request):
     method = request.method
     if method == "GET":
-        # TODO: only visible posts or something
         page, size, filter_ = get_post_query_params(request)
-        posts = apply_filter(request, filter_)
+        posts = filter(
+                lambda post: user_is_authorized(request.user, post),
+                get_public_posts())
+
         posts_pages = Paginator(posts, size)
         response_body = {
             "query": "posts",
@@ -98,7 +113,11 @@ def all_posts(request):
 def posts_by_pid(request, pid):
     method = request.method
     if method == "GET":
-        post = Post.objects.get(pk=pid)
+        post = get_object_or_404(Post, pk=pid)
+        if not user_is_authorized(request.user, post):
+            # returning forbidden would leak information
+            return HttpResponseNotFound()
+
         response_body = JSONRenderer().render({
             "query": "getPost",
             "post": PostSerializer(post).data
@@ -131,7 +150,11 @@ def comments_by_pid(request, pid):
     method = request.method
     if method == "GET":
         page, size, filter_ = get_post_query_params(request)
-        post = Post.objects.get(pk=pid)
+        post = get_object_or_404(Post, pk=pid)
+        if not user_is_authorized(request.user, post):
+            # returning forbidden would leak information
+            return HttpResponseNotFound()
+
         comments = Comment.objects.filter(post=post)
         comments_pages = Paginator(comments, size)
         response_body = {
@@ -159,7 +182,14 @@ def comments_by_pid(request, pid):
 # TODO: pid not actually needed, but we can check cid is a comment of pid if we want
 def comments_by_cid(request, pid, cid):
     method = request.method
-    comment = Comment.objects.get(pk=cid)
+
+    post = get_object_or_404(Post, pk=pid)
+    if not user_is_authorized(request.user, post):
+        # returning forbidden would leak information
+        return HttpResponseNotFound()
+
+    comment = get_object_or_404(post=post, pk=cid)
+
     if comment.author.id == request.user.pk:
         if method == "DELETE":
             comment.delete()
